husky_navigation:planner:Do not use quadratic and grid_path
*Using quadratic for our maps is producing strange offsets.our grid blocks has
got 1m side.quadratic is tending to stick the robot on the top-left corner of the block

*use_grid_path=false
Using this setting allows robot to freely move inside the grid blocks.
if its true,robot is forced to always pass through the middle of the block.
This situation is uncomfortable on our maps